requirements.txt
streamlit
cairosvg
gdown
pillow
plotly
lxml
pandas
matplotlib